+++ cmd/goose/driver_azuresql.go
@@ -1,0 +1,7 @@
+//go:build !no_mssql && !no_azuresql
+
+package main
+
+import (
+	_ "github.com/microsoft/go-mssqldb/azuread"
+)